machinelearning/rnnlm/requirements.txt
mecab-python3==0.7
joblib
keras
